Refactor code for better folder structure
