+++ tsconfig.json
@@ -3,8 +3,9 @@
     "target": "ES6",
     "module": "commonjs",
     "outDir": "./dist",
-    "strict": true
+    "strict": true,
+    "esModuleInterop": true
   },
-  "include": ["src/**/*.ts", "app.ts"],
+  "include": ["src/**/*.ts", "src/app.ts"],
   "exclude": ["node_modules"]
 }
